requirements.txt
Django==2.2.10
django-readonly-field==1.0.5
djangorestframework==3.12.2
pytz==2020.4
sqlparse==0.4.1
